+++ src/service/axios.config.js
@@ -1,5 +1,7 @@
 import axios from 'axios';
+import moment from 'moment';
 import { toast } from 'react-toastify';
+
 axios.defaults.baseURL = `http://localhost:5050/api/v1`;
 
 // axios.defaults.headers.common['Authorization'] = `Bearer ${localStorage.getItem(
@@ -50,5 +52,51 @@ export const current = async () => {
     return res;
   } catch (error) {
     toast.error('User not found');
+  }
+};
+
+
+//==================PRODUCTS ADD=====================
+export const addProduct = async product => {
+  try {
+    const { data } = await axios.post('/product/addDiaryFood', product);
+    return data;
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+//==================PRODUCTS DELETE=====================
+
+export const deleteProduct = async id => {
+  try {
+    const deletedProduct = await axios.delete("/product/delDiaryFood", { id });
+    return deletedProduct;
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+//==================PRODUCTS GET BY QUERY=====================
+
+export const getProductByQuery = async query => {
+  try {
+    const { data } = await axios.get(`/product/${query}`);
+    return data;
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+//==================PRODUCTS GET BY DATA=====================
+
+export const getProductsListByDate = async date => {
+  try {
+    const { data } = await axios.post("/users/dayinfo", {
+      day: moment(date).format('DD.MM.yyyy'),
+    });
+    return data;
+  } catch (error) {
+    console.log(error);
   }
 };
